[REFACT] ExerciseController 기능별 컨트롤러 분리 (#615)
* refactor: 운동 라이프사이클 컨트롤러 분리 - 운동 생성, 삭제, 수정, 상세 조회

* fix: 운동 라이프사이클 컨트롤러 반환 객체를 ResponseEntity로 통일

* fix: 운동 참여 컨트롤러 분리

* refactor: 운동 참여 서비스의 RequestMapping 공통으로 추출

* refactor: 운동 게스트 컨트롤러 분리

* docs: 스웨거 태그를 통일하여 분리되지 않도록 함

* refactor: 나와 관련된 운동 컨트롤러 분리

* refactor: 모임 운동 컨트롤러 분리

* refactor: 운동 추천 컨트롤러 분리

* refactor: 운동 맵 컨트롤러로 이름 변경

* chore: 빈 줄 제거

* refactor: Exercise Controller의 반환을 전부 ResponseEntity로 묶어서 반환하도록 변경

* chore: AGENTS.MD 파일 업데이트

* refactor: 게스트 컨트롤러 api 인터페이스 분리

* refactor: 라이프사이클 컨트롤러 api 인터페이스 분리

* refactor: Map 컨트롤러 api 인터페이스 분리

* refactor: My 컨트롤러 api 인터페이스 분리

* refactor: 참여 컨트롤러 api 인터페이스 분리

* refactor: 추천 컨트롤러 api 인터페이스 분리

* refactor: 파티 운동 컨트롤러 api 인터페이스 분리

* refactor: 운동 Api의 공통 태그를 어노테이션으로 분리

* refactor: Api 구현체인 컨트롤러의 파라미터는 가독성을 위해 개선
